requirements.txt
pydantic
pytest
pytest-cov
requests
python-dotenv
PyGithub
groq
pydantic-ai